requirements.txt
amqp==2.6.1
appdirs==1.4.3
apturl==0.5.2
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
billiard==3.6.3.0
blinker==1.4
Brlapi==0.7.0
CacheControl==0.12.6
celery==4.4.7
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==7.1.2
colorama==0.4.3
command-not-found==0.3
contextlib2==0.6.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-rest-auth==1.1.2
Django==3.0.8
django-allauth==0.42.0
django-rest-auth==0.9.5
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
install==1.3.3
ipaddr==2.2.0
isort==5.5.4
itsdangerous==1.1.0
Jinja2==2.11.2
kaleido==0.0.2
keyring==18.0.1
kombu==4.6.11
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
lxml==4.2.1
macaroonbakery==1.3.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
mysqlclient==2.0.1
netifaces==0.10.4
numexpr==2.7.1
numpy==1.19.1
oauthlib==3.1.0
olefile==0.46
orca==1.5.3
packaging==20.3
pandas==1.0.5
pdf2image==0.1.14
pdfkit==0.6.1
pep517==0.8.2
pexpect==4.6.0
Pillow==5.1.0
pipenv==2020.8.13
plotly==4.9.0
progress==1.5
protobuf==3.6.1
psutil==5.7.2
pycairo==1.16.2
pycodestyle==2.6.0
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
PyPDF2==1.26.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-decouple==3.3
python-pptx==0.6.10
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
retrying==1.3.3
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
systemd-python==234
tables==3.6.1
toml==0.10.1
toolz==0.10.0
ttystatus==0.38
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vboxapi==1.0
vine==1.3.0
virtualenv==20.0.32
virtualenv-clone==0.5.4
wadllib==1.3.3
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
xhtml2pdf==0.2.4
xkit==0.0.0
XlsxWriter==1.0.5
